% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalFlow.R
\name{calcAllChemicalFlow}
\alias{calcAllChemicalFlow}
\title{Aggregates chemical flow data in 2020 from calcAllChemicalRoute into broader categories
based on their outputs (ammonia, methanol, HVC, fertilizer and final ammonia/methanol).}
\usage{
calcAllChemicalFlow()
}
\description{
Aggregates chemical flow data in 2020 from calcAllChemicalRoute into broader categories
based on their outputs (ammonia, methanol, HVC, fertilizer and final ammonia/methanol).
}
\author{
Qianzhi Zhang
}
