% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalFlow2005_2020.R
\name{calcAllChemicalFlow2005_2020}
\alias{calcAllChemicalFlow2005_2020}
\title{Extrapolates Chemical Flows from 2020 to 2005-2020 based on total chemical UE in 2005-2020
and assuming constant UE shares of the final chemical flows;
intermediate ammonia and methanol flows are calculated from the final flows and 
ammonia-to-fertilizer and methanol-to-hvc ratios.
Flows are aggregated to the country level.}
\usage{
calcAllChemicalFlow2005_2020()
}
\description{
Extrapolates Chemical Flows from 2020 to 2005-2020 based on total chemical UE in 2005-2020
and assuming constant UE shares of the final chemical flows;
intermediate ammonia and methanol flows are calculated from the final flows and 
ammonia-to-fertilizer and methanol-to-hvc ratios.
Flows are aggregated to the country level.
}
\author{
Qianzhi Zhang
}
