% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalRoute2020.R
\name{calcAllChemicalRoute2020}
\alias{calcAllChemicalRoute2020}
\title{Filters chemical route flows from calcAllChemicalRoute2005_2020 for the year 2020
can be removed if calcAllChemicalRoute2005_2020 is read into REMIND}
\usage{
calcAllChemicalRoute2020(CCS = FALSE)
}
\arguments{
\item{CCS}{boolean parameter whether CCS technologies are considered as such in 2020 or assumed to be technologies without CCS}
}
\description{
Filters chemical route flows from calcAllChemicalRoute2005_2020 for the year 2020
can be removed if calcAllChemicalRoute2005_2020 is read into REMIND
}
\author{
Qianzhi Zhang
}
