% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalRoutes_2005to2020.R
\name{calcAllChemicalRoutes_2005to2020}
\alias{calcAllChemicalRoutes_2005to2020}
\title{Calculates chemical route flows (including OtherChem) from 2005 to 2020 
from total chemical flows based on 2020 route shares.
Flows are aggregated to the country level.}
\usage{
calcAllChemicalRoutes_2005to2020(CCS = FALSE)
}
\arguments{
\item{CCS}{boolean parameter whether CCS technologies are considered as such in 2020 or assumed to be technologies without CCS}
}
\description{
Calculates chemical route flows (including OtherChem) from 2005 to 2020 
from total chemical flows based on 2020 route shares.
Flows are aggregated to the country level.
}
\author{
Qianzhi Zhang
}
