% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalRoutes_2020.R
\name{calcAllChemicalRoutes_2020}
\alias{calcAllChemicalRoutes_2020}
\title{Filters chemical route flows from calcAllChemicalRoutes_2005to2020 for the year 2020
can be removed if calcAllChemicalRoutes_2005to2020 is read into REMIND}
\usage{
calcAllChemicalRoutes_2020(CCS = FALSE)
}
\arguments{
\item{CCS}{boolean parameter whether CCS technologies are considered as such in 2020 or assumed to be technologies without CCS}
}
\description{
Filters chemical route flows from calcAllChemicalRoutes_2005to2020 for the year 2020
can be removed if calcAllChemicalRoutes_2005to2020 is read into REMIND
}
\author{
Qianzhi Zhang
}
