% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalSpecFeDemand.R
\name{calcAllChemicalSpecFeDemand}
\alias{calcAllChemicalSpecFeDemand}
\title{Calculates chemical energy demand from 2005 to 2020 from chemical production per route (AllChemicalRoute2005_2020) 
and specific energy consumption for the different routes (retrieved from IEA_PetrochemEI and other literature sources).
The energy demand for OtherChem is calculated as the remaining share of total chemical industry energy demand. 
Results are aggregated to the country level.}
\usage{
calcAllChemicalSpecFeDemand(CCS = FALSE)
}
\arguments{
\item{CCS}{boolean parameter whether CCS technologies are considered as such in 2020 or assumed to be technologies without CCS}
}
\description{
Calculates chemical energy demand from 2005 to 2020 from chemical production per route (AllChemicalRoute2005_2020) 
and specific energy consumption for the different routes (retrieved from IEA_PetrochemEI and other literature sources).
The energy demand for OtherChem is calculated as the remaining share of total chemical industry energy demand. 
Results are aggregated to the country level.
}
\author{
Qianzhi Zhang
}
