% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalUe.R
\name{calcAllChemicalUe}
\alias{calcAllChemicalUe}
\title{Calculates UE shares for all chemicals (including OtherChem) in 2020 from 
final product flows (methFinal, ammoFinal, HVC, fertilizer), 
material prices (mat2ue) and the total UE of chemicals (from FeDemandIndustry).
Shares are aggregated to the country level.}
\usage{
calcAllChemicalUe()
}
\description{
Calculates UE shares for all chemicals (including OtherChem) in 2020 from 
final product flows (methFinal, ammoFinal, HVC, fertilizer), 
material prices (mat2ue) and the total UE of chemicals (from FeDemandIndustry).
Shares are aggregated to the country level.
}
\author{
Qianzhi Zhang
}
