% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalUeShare.R
\name{calcAllChemicalUeShare}
\alias{calcAllChemicalUeShare}
\title{Calculates shares of ammoFinal, methFinal, HVC, fertilizer and OtherChem
of total chemical UE for 2020-2050 based on the UE shares in 2020 and the 
projected relative increases in production (IEA The Future of Petrochemicals) and 
total chemical UE (FeDemandIndustry)}
\usage{
calcAllChemicalUeShare()
}
\description{
Calculates shares of ammoFinal, methFinal, HVC, fertilizer and OtherChem
of total chemical UE for 2020-2050 based on the UE shares in 2020 and the 
projected relative increases in production (IEA The Future of Petrochemicals) and 
total chemical UE (FeDemandIndustry)
}
\author{
Qianzhi Zhang
}
