% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllChemicalUeShares_2020to2150.R
\name{calcAllChemicalUeShares_2020to2150}
\alias{calcAllChemicalUeShares_2020to2150}
\title{Calculates shares of ammoFinal, methFinal, HVC, fertilizer and OtherChem
of total chemical UE for 2020-2150 based on the UE shares in 2020 and the
projected relative increases in production (IEA The Future of Petrochemicals) and
total chemical UE (FeDemandIndustry)}
\usage{
calcAllChemicalUeShares_2020to2150()
}
\description{
Calculates shares of ammoFinal, methFinal, HVC, fertilizer and OtherChem
of total chemical UE for 2020-2150 based on the UE shares in 2020 and the
projected relative increases in production (IEA The Future of Petrochemicals) and
total chemical UE (FeDemandIndustry)
}
\author{
Qianzhi Zhang
}
