% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAmmoniaRoute.R
\name{calcAmmoniaRoute}
\alias{calcAmmoniaRoute}
\title{Calculates ammonia production volumes per production route in Mt for 2015-2020
based on IFA total ammonia production volumes (2015-2020) and
shares for different ammonia production routes from IEA (2020 and 2017) and China specific data (2020)}
\usage{
calcAmmoniaRoute()
}
\description{
Calculates ammonia production volumes per production route in Mt for 2015-2020
based on IFA total ammonia production volumes (2015-2020) and
shares for different ammonia production routes from IEA (2020 and 2017) and China specific data (2020)
}
\author{
Qianzhi Zhang
}
