% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBinderProduction.R
\name{calcBinderProduction}
\alias{calcBinderProduction}
\title{Calculates global cement production as from Andrew's 2019 paper.}
\usage{
calcBinderProduction(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement or "clinker".}
}
\description{
Calculates global cement production as from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
