% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBuiltLifespan.R
\name{calcBuiltLifespan}
\alias{calcBuiltLifespan}
\title{Calculates the lifetimes of residential and non-residential buildings, as well as of civil engeneering.}
\usage{
calcBuiltLifespan()
}
\description{
Calculates the lifetimes of residential and non-residential buildings, as well as of civil engeneering.
}
\author{
Bennet Weiss
}
