% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCementProduction.R
\name{calcCementProduction}
\alias{calcCementProduction}
\title{Calculates global cement production as from Andrew's 2019 paper.}
\usage{
calcCementProduction(x)
}
\arguments{
\item{x}{magclass object}
}
\description{
Calculates global cement production as from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
