% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChemicalFeedstocks.R
\name{calcChemicalFeedstocks}
\alias{calcChemicalFeedstocks}
\title{Calculate different feedstocks for the chemical sector from IEA energy balances}
\usage{
calcChemicalFeedstocks()
}
\description{
Calculate different feedstocks for the chemical sector from IEA energy balances
}
\author{
Qianzhi Zhang
}
