% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChemicalFlows_2020.R
\name{calcChemicalFlows_2020}
\alias{calcChemicalFlows_2020}
\title{Aggregates chemical flow data in 2020 from calcAllChemicalRoutes_2020into broader categories
based on their outputs (ammonia, methanol, HVC, fertilizer and final ammonia/methanol).}
\usage{
calcChemicalFlows_2020()
}
\description{
Aggregates chemical flow data in 2020 from calcAllChemicalRoutes_2020into broader categories
based on their outputs (ammonia, methanol, HVC, fertilizer and final ammonia/methanol).
}
\author{
Qianzhi Zhang
}
