% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChemicalRoutes_2020.R
\name{calcChemicalRoutes_2020}
\alias{calcChemicalRoutes_2020}
\title{Joins magpie objects with ammonia, methanol, HVC (Mt) and fertilizer production routes (MtN),
calculates ammonia and methanol final routes (Mt that are not used for HVC/fertilizer production),
converts from Mt to Gt and returns magpie object for year 2020.}
\usage{
calcChemicalRoutes_2020()
}
\description{
Joins magpie objects with ammonia, methanol, HVC (Mt) and fertilizer production routes (MtN),
calculates ammonia and methanol final routes (Mt that are not used for HVC/fertilizer production),
converts from Mt to Gt and returns magpie object for year 2020.
}
\author{
Qianzhi Zhang
}
