% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChemicalTotal.R
\name{calcChemicalTotal}
\alias{calcChemicalTotal}
\title{Calculate Total Chemical Energy consumption from IEA energy balances}
\usage{
calcChemicalTotal()
}
\description{
Calculate Total Chemical Energy consumption from IEA energy balances
}
\author{
Qianzhi Zhang
}
