% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChinaBaogao.R
\name{calcChinaBaogao}
\alias{calcChinaBaogao}
\title{Retrieve Methanol capacity, Production and consumption data in China from ChinaBaogao 
(Chinese business website that publishes industry analyses.)}
\usage{
calcChinaBaogao()
}
\description{
Retrieve Methanol capacity, Production and consumption data in China from ChinaBaogao 
(Chinese business website that publishes industry analyses.)
}
\author{
Qianzhi Zhang
}
