% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClinkerRatio.R
\name{calcClinkerRatio}
\alias{calcClinkerRatio}
\title{Calculates global clinker ratio by employing cement and clinker production from Andrew's 2019 paper.}
\usage{
calcClinkerRatio()
}
\description{
Calculates global clinker ratio by employing cement and clinker production from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
