% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClinkerToCementRatio.R
\name{calcClinkerToCementRatio}
\alias{calcClinkerToCementRatio}
\title{Calculate Clinker-to-Cement Ratio}
\usage{
calcClinkerToCementRatio()
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, and \code{description}.
}
\description{
Calculate Clinker-to-Cement Ratio
}
\seealso{
\code{\link[=readADVANCE_WP2]{readADVANCE_WP2()}}, \code{\link[=convertADVANCE_WP2]{convertADVANCE_WP2()}}
}
\author{
Michaja Pehl
}
