% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCullenStaticParameters.R
\name{calcCullenStaticParameters}
\alias{calcCullenStaticParameters}
\title{Calc Cullen Static Parameters}
\usage{
calcCullenStaticParameters(subtype)
}
\arguments{
\item{subtype}{Subtype of static parameter to calculate. Currently supported 
subtypes are:
'productionLossRate', 'formingLossRate', 'formingYield', 'fabricationYield'.}
}
\description{
Calc static (singular) steel parameters based on supplemnetar information of
Cullen et al. (2012) paper
'Mapping the global flow of steel: from steelmaking to end‐use goods'. See
\link[readCullen2012]{readCullen2012} and the folder Cullen2012 in sources
for details on the preprocessing.
}
\author{
Merlin Jo Hosak
}
