% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDRIData.R
\name{calcDRIData}
\alias{calcDRIData}
\title{Calc DRI data}
\usage{
calcDRIData(subtype)
}
\arguments{
\item{subtype}{Options: 'production', 'imports', 'exports'}
}
\description{
Cald direct reduced iron data mostly based on World Steel Association
data (see |link[readWorldSteelDatabase]{readWorldSteelDatabase} for details).
The DRI data is backcasted based on steel production data to fill in missing
years.
}
\author{
Merlin Jo Hosak
}
