% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyBalancesOutputToIndustry.R
\name{calcEnergyBalancesOutputToIndustry}
\alias{calcEnergyBalancesOutputToIndustry}
\title{Calculate Energy Balances Output to Industry}
\usage{
calcEnergyBalancesOutputToIndustry()
}
\description{
Calculate Energy Balances Output to Industry
}
\author{
Michaja Pehl, Falk Benke
}
