% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFertilizerRoute.R
\name{calcFertilizerRoute}
\alias{calcFertilizerRoute}
\title{Calculates the total fertilizer production (MtN, as the sum of urea, AN, CAN and AS) per country 
and the fraction of ammonia production (MtN/MtN) used for fertilizer production based on IFA data.}
\usage{
calcFertilizerRoute()
}
\description{
Calculates the total fertilizer production (MtN, as the sum of urea, AN, CAN and AS) per country 
and the fraction of ammonia production (MtN/MtN) used for fertilizer production based on IFA data.
}
\author{
Qianzhi Zhang
}
