% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDP200.R
\name{calcGDP200}
\alias{calcGDP200}
\title{Get GDP from 1900-2100}
\usage{
calcGDP200(scenario = "SSP2", per_capita = FALSE)
}
\arguments{
\item{scenario}{Scenario to use for future GDP data (default: SSP2).}

\item{per_capita}{If TRUE, GDP is returned as per capita (default: FALSE).}
}
\value{
List with Magpie object of GDP and metadata in calcOutput format.
}
\description{
Calc GDP (PPP) from 1900-2100 yearly for the SIMSON format on a country 
level. Can be aggregated to regions via calcOutput aggregate parameter.
Uses \link[=readOECD_GDP]{OECD_GDP} for historical GDP data as well as
\link[=calcGDP]{calcGDP} from mrdrivers for current and future GDP data
according to a specific scenario (see \code{vignette("scenarios")} 
for more information). Population data from 
\link[=calcPopulation200]{calcPopulation200} is used to convert 
GDP per capita to total GDP.
GDP is given in 2005 USD (PPP). It's extrapolated to the past with historic
GDP datasets that use a different base year, which however does not matter
as only the relative values are used 
(see \link[=toolInterpolate]{toolInterpolate}).
}
\author{
Merlin Jo Hosak
}
