% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistoricalBasicMaterialProduction.R
\name{calcHistoricalBasicMaterialProduction}
\alias{calcHistoricalBasicMaterialProduction}
\title{Calculate historical basic material production data for
steel (Worldsteel) and cement (USGS)}
\usage{
calcHistoricalBasicMaterialProduction(subtype)
}
\arguments{
\item{subtype}{Either 'cement' or 'steel'}
}
\description{
Calculate historical basic material production data for
steel (Worldsteel) and cement (USGS)
}
\author{
Michaja Pehl, Falk Benke
}
