% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistoricalProduction.R
\name{calcHistoricalProduction}
\alias{calcHistoricalProduction}
\title{Calculate historical production data for steel (Worldsteel) and cement (USGS)}
\usage{
calcHistoricalProduction(subtype)
}
\arguments{
\item{subtype}{Either 'cement' or 'steel'}
}
\description{
Calculate historical production data for steel (Worldsteel) and cement (USGS)
}
\author{
Michaja Pehl, Falk Benke
}
