% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistoricalSteelStock.R
\name{calcHistoricalSteelStock}
\alias{calcHistoricalSteelStock}
\title{Calculate Steel Stock from Mueller steel stock per capita and WDI population}
\usage{
calcHistoricalSteelStock()
}
\description{
Calculate Steel Stock from Mueller steel stock per capita and WDI population
}
\author{
Falk Benke
}
