% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHVCRoute.R
\name{calcHVCRoute}
\alias{calcHVCRoute}
\title{Calculates HVC production volumes per production route in Mt for 2015-2020
based on IEA The Future of Petrochemicals (2018)}
\usage{
calcHVCRoute()
}
\description{
Calculates HVC production volumes per production route in Mt for 2015-2020
based on IEA The Future of Petrochemicals (2018)
}
\author{
Qianzhi Zhang
}
