% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_Ammonia.R
\name{calcIEA_Ammonia}
\alias{calcIEA_Ammonia}
\title{Read-in IEA Ammonia Technology Roadmap 2021 Fig 2.9 Ammonia production by process route and
scenario in major ammonia producing regions data as a magclass object.}
\usage{
calcIEA_Ammonia(subtype)
}
\arguments{
\item{subtype}{Different scenarios of Ammonia data that should be read. Available types are:
\itemize{
  \item BaseYear_2020: Base year data in 2020
  \item STEPS_2050: IEA STEPS scenario in 2050
  \item SDS_2050: IEA SDS scenario in 2050
}}
}
\description{
Read-in IEA Ammonia Technology Roadmap 2021 Fig 2.9 Ammonia production by process route and
scenario in major ammonia producing regions data as a magclass object.
}
\author{
Qianzhi Zhang
}
