% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEA_PetrochemEI.R
\name{calcIEA_PetrochemEI}
\alias{calcIEA_PetrochemEI}
\title{Retrieve specific energy consumption (SEC) for the production of key chemicals (2006)
from IEA Information Paper Chemical and Petrochemical Sector 2009 (Table 12)}
\usage{
calcIEA_PetrochemEI()
}
\description{
Retrieve specific energy consumption (SEC) for the production of key chemicals (2006)
from IEA Information Paper Chemical and Petrochemical Sector 2009 (Table 12)
}
\author{
Qianzhi Zhang
}
