% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFA_Chem.R
\name{calcIFA_Chem}
\alias{calcIFA_Chem}
\title{Retrieve IFA (International fertilizer Association) data containing 
production, consumption, export and import volumes as well as capacities for 
ammonia and urea.}
\usage{
calcIFA_Chem(subtype, unitNitrogen = FALSE)
}
\arguments{
\item{subtype}{Character string specifying the type of IFA product data to read.
Available types are:
\itemize{
  \item For product: ammonia, urea
  \item For data sheet: statistics, capacities
  \item For product characteristics: consumption, production, export, import, capacities
}}

\item{unitNitrogen}{boolean parameter
TRUE to return data in unit KtN 
FALSE to return data in Kt product}
}
\value{
magpie object of the IFA data
}
\description{
Retrieve IFA (International fertilizer Association) data containing 
production, consumption, export and import volumes as well as capacities for 
ammonia and urea.
}
\author{
Qianzhi Zhang
}
