% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFA_ChemAppend.R
\name{calcIFA_ChemAppend}
\alias{calcIFA_ChemAppend}
\title{Retrieves IFA (International fertilizer Association) data containing
production volumes and/or capacities for ammonium nitrate (AN), ammonium sulphate (AS),
calcium ammonium nitrate (CAN) and urea ammonium nitrate (UAN).}
\usage{
calcIFA_ChemAppend(subtype, unitNitrogen = FALSE)
}
\arguments{
\item{subtype}{Character string indicating the IFA product and data type to read.
Available combinations include:
\itemize{
  \item AN_statistics_production
  \item AS_statistics_production
  \item CAN_statistics_production
  \item AN_capacities_capacities, AS_capacities_capacities, UAN_capacities_capacities
}}

\item{unitNitrogen}{boolean parameter
TRUE to return production statistics in unit KtN 
FALSE to return production statistics in Kt product
capacities are always in KtN}
}
\value{
Magpie object of the IFA data.
}
\description{
Retrieves IFA (International fertilizer Association) data containing
production volumes and/or capacities for ammonium nitrate (AN), ammonium sulphate (AS),
calcium ammonium nitrate (CAN) and urea ammonium nitrate (UAN).
}
\author{
Qianzhi Zhang
}
