% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIHS_Meth.R
\name{calcIHS_Meth}
\alias{calcIHS_Meth}
\title{Retrieve methanol production, capacity and demand data for 2018 from 
9TH RUSSIA & CIS OIL & GAS EXECUTIVE SUMMIT 2019 METHANOL,
optionally the 2018 data is interpolated to 2010-2020 based on 
temporal coverage of Methanol capacities 2010-2020 from IHS Markit}
\usage{
calcIHS_Meth(subtype)
}
\arguments{
\item{subtype[1]}{Type of Methanol data to read. Available types are:
\itemize{
  \item Production
  \item Capacity
  \item Demand
}}

\item{subtype[2]}{Temporal coverage of Methanol data. Available types are:
\itemize{
  \item 2010-2020: Interpolated 2018 data to 2010-2020 using total global methanol production data.
  \item 2018: Data for the year 2018.
}}
}
\value{
Magpie object of the IHS_Meth data.
}
\description{
Retrieve methanol production, capacity and demand data for 2018 from 
9TH RUSSIA & CIS OIL & GAS EXECUTIVE SUMMIT 2019 METHANOL,
optionally the 2018 data is interpolated to 2010-2020 based on 
temporal coverage of Methanol capacities 2010-2020 from IHS Markit
}
\author{
Qianzhi Zhang
}
