% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustryFE.R
\name{calcIndustryFE}
\alias{calcIndustryFE}
\title{Calculate different feedstocks for the chemical sector from IEA energy balances}
\usage{
calcIndustryFE()
}
\description{
Calculate different feedstocks for the chemical sector from IEA energy balances
}
\author{
Qianzhi Zhang
}
