% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustrySubsectorsOutput.R
\name{calcIndustrySubsectorsOutput}
\alias{calcIndustrySubsectorsOutput}
\title{Calculate Industry Subsectors Output
Additional corrections are applied to the IEA data in [`mrindustry::tool_fix_IEA_data_for_Industry_subsectors`].}
\usage{
calcIndustrySubsectorsOutput()
}
\description{
Calculate Industry Subsectors Output
Additional corrections are applied to the IEA data in [`mrindustry::tool_fix_IEA_data_for_Industry_subsectors`].
}
\author{
Michaja Pehl, Falk Benke
}
