% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMaterialTrade.R
\name{calcMaterialTrade}
\alias{calcMaterialTrade}
\title{Calculates global clinker or cement trade (net import)
based on Chatham House, UN Comtrade, and USGS (in the case of cement).}
\usage{
calcMaterialTrade(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement" or "clinker".}
}
\description{
Calculates global clinker or cement trade (net import)
based on Chatham House, UN Comtrade, and USGS (in the case of cement).
}
\author{
Bennet Weiss
}
