% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMCeBinderProduction.R
\name{calcMCeBinderProduction}
\alias{calcMCeBinderProduction}
\title{Calculates global cement production as from Andrew's 2019 paper.}
\usage{
calcMCeBinderProduction(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement or "clinker".}
}
\description{
Calculates global cement production as from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
