% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMCeBuiltLifespan.R
\name{calcMCeBuiltLifespan}
\alias{calcMCeBuiltLifespan}
\title{Calculates the lifetimes of residential and non-residential buildings, as well as of civil engeneering.}
\usage{
calcMCeBuiltLifespan()
}
\description{
Calculates the lifetimes of residential and non-residential buildings, as well as of civil engeneering.
}
\author{
Bennet Weiss
}
