% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMCeClinkerRatio.R
\name{calcMCeClinkerRatio}
\alias{calcMCeClinkerRatio}
\title{Calculates global clinker ratio by employing cement and clinker production from Andrew's 2019 paper.}
\usage{
calcMCeClinkerRatio()
}
\description{
Calculates global clinker ratio by employing cement and clinker production from Andrew's 2019 paper.
}
\author{
Bennet Weiss
}
