% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMMSA_Methanol.R
\name{calcMMSA_Methanol}
\alias{calcMMSA_Methanol}
\title{Read-in MMSA Global Methanol Outlook 2023 Growth and Decarbonization data
containing regional methanol capacities and demands.}
\usage{
calcMMSA_Methanol()
}
\description{
Read-in MMSA Global Methanol Outlook 2023 Growth and Decarbonization data
containing regional methanol capacities and demands.
}
\author{
Qianzhi Zhang
}
