% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMPlConsumption.R
\name{calcMPlConsumption}
\alias{calcMPlConsumption}
\title{Calculate Country-Level Total Plastic Use}
\usage{
calcMPlConsumption()
}
\description{
Load OECD regional plastic use data and trade/production inputs,
adjust for domestic demand, normalize baselines, and aggregate to country level.
}
\author{
Qianzhi Zhang
}
