% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMPlLifetime.R
\name{calcMPlLifetime}
\alias{calcMPlLifetime}
\title{Calculates the mean lifetimes of plastic goods by use sector, or their standard deviations.}
\usage{
calcMPlLifetime(subtype)
}
\arguments{
\item{subtype}{Character string specifying to read means or standard deviations from data
- "Lifetime_mean"
- "Lifetime_std"}
}
\description{
Calculates the mean lifetimes of plastic goods by use sector, or their standard deviations.
}
\author{
Leonie Schweiger
}
