% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMPlOECD_MGshare.R
\name{calcMPlOECD_MGshare}
\alias{calcMPlOECD_MGshare}
\title{Calculate Country-Level Material Good Shares}
\usage{
calcMPlOECD_MGshare()
}
\description{
Compute material share of different goods from 2019 OECD use data,
aggregate regional shares to country level for 2019.
}
\author{
Qianzhi Zhang
}
