% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_Plastic.R
\name{calcOECD_Plastic}
\alias{calcOECD_Plastic}
\title{This function returns OECD plastic use or waste data from OECD Plastic outlook
based on a specified subtype in a MagPIE object aggregated to ISO country level.}
\usage{
calcOECD_Plastic(subtype)
}
\arguments{
\item{subtype}{Character string specifying the dataset and scope.
Valid formats include:
- "Use_2019_region" (plastic use in Mt, dimensions: region, application, polymer)
- "Use_1990-2019_region" (plastic use in Mt, dimensions: region, year)
- "Use_1990-2019_world" (plastic use in Mt, dimensions: year, application, polymer, type (primary/secondary))
- "WasteEOL_1990-2019_region" (plastic waste in Mt, dimensions: region, year, EOL fate, collected for recycling)
- "WasteType_2019_region" (plastic waste in Mt, dimensions: region, application, polymer)}
}
\description{
This function returns OECD plastic use or waste data from OECD Plastic outlook
based on a specified subtype in a MagPIE object aggregated to ISO country level.
}
\author{
Qianzhi Zhang
}
