% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticChemRe.R
\name{calcOECD_PlasticChemRe}
\alias{calcOECD_PlasticChemRe}
\title{Calculate Country-Level Chemical Recycling Rate Trajectories}
\usage{
calcOECD_PlasticChemRe(target = NULL)
}
\description{
Generate time series of chemical recycling share trajectories by sector and region,
then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
