% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticCollRate.R
\name{calcOECD_PlasticCollRate}
\alias{calcOECD_PlasticCollRate}
\title{Calculate Country-Level Plastic Collection Rate Trajectories}
\usage{
calcOECD_PlasticCollRate()
}
\description{
Build time series of plastic collection rates by sector and region,
apply fixed and interpolated values, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
