% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticDACRate.R
\name{calcOECD_PlasticDACRate}
\alias{calcOECD_PlasticDACRate}
\title{Calculate Country-Level DAC-based Plastic Share Trajectories}
\usage{
calcOECD_PlasticDACRate()
}
\description{
Generate time series of Direct-Air-Capture-based plastic share trajectories by sector
and aggregate from regions to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
