% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticEoL.R
\name{calcOECD_PlasticEoL}
\alias{calcOECD_PlasticEoL}
\title{Calculate Country-Level End-of-Life Plastic Fate Shares}
\usage{
calcOECD_PlasticEoL()
}
\description{
Compute end-of-life fate ratios of plastics by country,
based on OECD regional waste EOL data (1990–2019).
}
\author{
Qianzhi Zhang
}
