% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticIncinRate.R
\name{calcOECD_PlasticIncinRate}
\alias{calcOECD_PlasticIncinRate}
\title{Calculate Country-Level Plastic Incineration Rate Trajectories}
\usage{
calcOECD_PlasticIncinRate()
}
\description{
Build time series of plastic incineration rates by sector and region
using OECD EoL data and external sources, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
