% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticLandfillRe.R
\name{calcOECD_PlasticLandfillRe}
\alias{calcOECD_PlasticLandfillRe}
\title{Calculate Country-Level Plastic Landfill Rate Trajectories}
\usage{
calcOECD_PlasticLandfillRe()
}
\description{
Build time series of plastic landfill rates by sector and region
using OECD end-of-life (EoL) data, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
