% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticMechLoss.R
\name{calcOECD_PlasticMechLoss}
\alias{calcOECD_PlasticMechLoss}
\title{Calculate Country-Level Mechanical Recycling Loss Trajectories}
\usage{
calcOECD_PlasticMechLoss()
}
\description{
Generate time series of mechanical recycling loss trajectories by sector and region,
then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
