% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticMechReRate.R
\name{calcOECD_PlasticMechReRate}
\alias{calcOECD_PlasticMechReRate}
\title{Calculate Country-Level Mechanical Recycling Rate Trajectories}
\usage{
calcOECD_PlasticMechReRate()
}
\description{
Generate time series of mechanical recycling rates by sector and region
using OECD end-of-life (EoL) data and external sources, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
