% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticMGshare.R
\name{calcOECD_PlasticMGshare}
\alias{calcOECD_PlasticMGshare}
\title{Calculate Country-Level Material Good Shares}
\usage{
calcOECD_PlasticMGshare()
}
\description{
Compute material share of different goods from 2019 use data,
aggregate regional shares to country level for 2019.
}
\author{
Qianzhi Zhang
}
