% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticUse.R
\name{calcOECD_PlasticUse}
\alias{calcOECD_PlasticUse}
\title{Calculate Country-Level Plastic Use by Sector}
\usage{
calcOECD_PlasticUse()
}
\description{
Combine sectoral use shares and total use to compute absolute plastic use
by sector at country level for 1990-2019.
}
\author{
Qianzhi Zhang
}
