% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticUseShare.R
\name{calcOECD_PlasticUseShare}
\alias{calcOECD_PlasticUseShare}
\title{Calculate OECD Plastic Use Share by Country}
\usage{
calcOECD_PlasticUseShare()
}
\description{
Compute the share of plastic use by sector for each country, based on regional OECD data
and supplemental EU shares, then aggregate to country level.
}
\author{
Qianzhi Zhang
}
