% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOECD_PlasticUseTotal.R
\name{calcOECD_PlasticUseTotal}
\alias{calcOECD_PlasticUseTotal}
\title{Calculate Country-Level Total Plastic Use}
\usage{
calcOECD_PlasticUseTotal()
}
\description{
Load OECD regional plastic use data and trade/production inputs,
adjust for domestic demand, normalize baselines, and aggregate to country level.
}
\author{
Qianzhi Zhang
}
