% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPigIronPreliminaryData.R
\name{calcPigIronPreliminaryData}
\alias{calcPigIronPreliminaryData}
\title{Calc Pig Iron Preliminary Data}
\usage{
calcPigIronPreliminaryData(subtype)
}
\arguments{
\item{subtype}{Options: 'production', 'imports', 'exports'}
}
\description{
Prepare pig iron data for final output functions. Loads World Steel Association
(WSA, s. \link[readWorldSteelDatabase]{readWorldSteelDatabase}) and 
Industrial Ecology Data Commons (IEDC, s. \link[readIEDC]{readIEDC}) data.
The WSA data is used as a baseline (as for many other steel related data)
and forecasted using IEDC data as more recent WSA data is missing.
}
\author{
Merlin Jo Hosak
}
