% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPigIronProduction.R
\name{calcPigIronProduction}
\alias{calcPigIronProduction}
\title{Calc pig iron production}
\usage{
calcPigIronProduction()
}
\description{
Calculate pig iron production. Wrapper function. See 
\link[calcPigIronPreliminaryData]{calcPigIronPreliminaryData} for specifics.
}
\author{
Merlin Jo Hosak
}
