% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPigIronTrade.R
\name{calcPigIronTrade}
\alias{calcPigIronTrade}
\title{Calc pig iron trade}
\usage{
calcPigIronTrade(subtype)
}
\arguments{
\item{subtype}{'imports' or 'exports' to calculate pig iron imports or 
exports respectively.}
}
\description{
Calculate pig iron trade Wrapper function. See 
\link[calcPigIronPreliminaryData]{calcPigIronPreliminaryData} for specifics.
}
\author{
Merlin Jo Hosak
}
