% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticBioRate.R
\name{calcPlasticBioRate}
\alias{calcPlasticBioRate}
\title{Calculate Country-Level Bio-based Plastic Share Trajectories}
\usage{
calcPlasticBioRate()
}
\description{
Generate time series of bio-based plastic share trajectories by sector
and aggregate from regions to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
