% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticChemRe.R
\name{calcPlasticChemRe}
\alias{calcPlasticChemRe}
\title{Calculate Country-Level Chemical Recycling Rate Trajectories}
\usage{
calcPlasticChemRe(target = NULL)
}
\description{
Generate time series of chemical recycling share trajectories by sector and region,
then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
