% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticCollRate.R
\name{calcPlasticCollRate}
\alias{calcPlasticCollRate}
\title{Calculate Country-Level Plastic Collection Rate Trajectories}
\usage{
calcPlasticCollRate()
}
\description{
Build time series of plastic collection rates by sector and region,
apply fixed and interpolated values, then aggregate to countries for 1990–2100.
}
\author{
Qianzhi Zhang
}
