% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticUse.R
\name{calcPlasticUse}
\alias{calcPlasticUse}
\title{Calculate Country-Level Plastic Use by Sector}
\usage{
calcPlasticUse()
}
\description{
Combine sectoral use shares and total use to compute absolute plastic use
by sector at country level for 1990-2019.
}
\author{
Qianzhi Zhang
}
