% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlasticUseShare.R
\name{calcPlasticUseShare}
\alias{calcPlasticUseShare}
\title{Calculate Plastic Use Share by Country}
\usage{
calcPlasticUseShare()
}
\description{
Compute the share of plastic use by sector for each country in 2019, 
based on regional OECD data and supplemental EU shares.
}
\author{
Qianzhi Zhang
}
