% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRMI_China.R
\name{calcRMI_China}
\alias{calcRMI_China}
\title{Read-in RMI (Rocky Mountain Institute) "Transforming China Chemicals Industry Pathways and Outlook under the Carbon Neutrality Goal 2022.xlsx"
data from either "ES1-3 China Chemical Demand" or "ES29 China Chemical Structure" sheets as a magclass object.
Data contains chemical demand projections 2020-2050 for ammonia, methanol and ethylene and the feedstock structure for the production of ammonia, methanol and ethylene
in the zero-carbon scenario.}
\usage{
calcRMI_China(subtype)
}
\arguments{
\item{subtype[1]}{Type of RMI_China data sheet to read. Available types are:
\itemize{
  \item ChemDemand: ES1-3 China Chemical Demand
  \item ChemStructure: ES10 China Chemical Structure
}}

\item{subtype[2]}{The specific product from the RMI_China data to read. Available types are:
\itemize{
  \item Ammonia
  \item Methanol
  \item Ethylene
}}
}
\description{
Read-in RMI (Rocky Mountain Institute) "Transforming China Chemicals Industry Pathways and Outlook under the Carbon Neutrality Goal 2022.xlsx"
data from either "ES1-3 China Chemical Demand" or "ES29 China Chemical Structure" sheets as a magclass object.
Data contains chemical demand projections 2020-2050 for ammonia, methanol and ethylene and the feedstock structure for the production of ammonia, methanol and ethylene
in the zero-carbon scenario.
}
\author{
Qianzhi Zhang
}
