% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteel_Projections.R
\name{calcSteel_Projections}
\alias{calcSteel_Projections}
\title{Function for calculating industry activity trajectories.}
\usage{
calcSteel_Projections(
  subtype = "production",
  scenarios,
  match.steel.historic.values = TRUE,
  match.steel.estimates = "none",
  save.plots = NULL,
  do_use_expert_guess = TRUE
)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{production} Returns trajectories of primary and secondary steel
production.
\item \code{secondary.steel.max.share} Returns the maximum share of secondary steel
in total steel production.
}}

\item{scenarios}{Vector of strings designating the scenarios to be returned.}

\item{match.steel.historic.values}{Should steel production trajectories match
historic values?}

\item{match.steel.estimates}{Should steel production trajectories match
exogenous estimates?
\itemize{
\item \code{none} no matching
\item \code{IEA_ETP} IEA 2017 Energy Transition Pathways steel production totals for
OECD and Non-OECD countries from the \emph{Reference Technologies Scenario}
until 2060, and original growth rates after that.
}}

\item{save.plots}{\code{NULL} (default) if no plots are saved, or the path to save
directories to.}

\item{do_use_expert_guess}{Whether or not to overwrite steel productions with
expert guesses from input data in the sources folder.}
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, \code{min}, and \code{max}.
}
\description{
Function for calculating industry activity trajectories.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
