% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelLifetimes.R
\name{calcSteelLifetimes}
\alias{calcSteelLifetimes}
\title{Calc Steel Lifetimes}
\usage{
calcSteelLifetimes(subtype)
}
\arguments{
\item{subtype}{Subtype of steel lifetimes data to load. Currently only 
'Cooper2014' is available. Other options might include Pauliuk or Wittig (
see respective folders).}
}
\description{
Function to load steel lifetimes data in years.
}
\author{
Merlin Jo Hosak
}
