% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelRecoveryRate.R
\name{calcSteelRecoveryRate}
\alias{calcSteelRecoveryRate}
\title{Calc steel recovery rate}
\usage{
calcSteelRecoveryRate(subtype)
}
\arguments{
\item{subtype}{Subtype of steel recovery rate data to load. Currently only
'WorldSteel' is available, see 
\link[readWorldSteelParameters]{readWorldSteelParameters} for details.}
}
\description{
Function to load steel recovery rate data in the four main end-use sectors
(construction, machinery, products, transport).
}
\author{
Merlin Jo Hosak
}
