% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteelTrade.R
\name{calcSteelTrade}
\alias{calcSteelTrade}
\title{Get steel trade data}
\usage{
calcSteelTrade(subtype = "imports")
}
\value{
Steel trade across all regions from 1900-2022 as magpie within 
list of metadata (in calcOutput format).
}
\description{
Calc steel trade from WorldSteel datasets. Can be aggregated to regions 
via calcOutput aggregate parameter. Uses 
\link[readWorldSteelDigitised]{WorldSteelDigitised} and 
\link[readWorldSteelDatabase]{WorldSteelDatabase} datasets, the former for
historic, the latter for current data. Further, uses 
\link[calcSteelProduction]{SteelProduction} to backcast historic trade data.
}
\author{
Merlin Jo Hosak
}
