% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWorldSteelStaticParameters.R
\name{calcWorldSteelStaticParameters}
\alias{calcWorldSteelStaticParameters}
\title{Calc World Steel Static Parameters}
\usage{
calcWorldSteelStaticParameters(subtype)
}
\arguments{
\item{subtype}{Parameter to load. Currently only 'scrapInBOFRate' available.}
}
\description{
Load static (singular) parameters based on World Steel Association data.
See \link[readWorldSteelParameters]{readWorldSteelParameters} and the 
WorldSteelParameters folder for more information.
}
\author{
Merlin Jo Hosak
}
