% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertChathamHouse.R
\name{convertChathamHouse}
\alias{convertChathamHouse}
\title{Convert trade data from Chatham House}
\usage{
convertChathamHouse(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert trade data from Chatham House
}
\author{
Bennet Weiss
}
