% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_Ammonia.R
\name{convertIEA_Ammonia}
\alias{convertIEA_Ammonia}
\title{Convert IEA_Ammonia}
\usage{
convertIEA_Ammonia(x)
}
\arguments{
\item{x}{MagPIE object containing IEA Ammonia data at regional resolution.}
}
\value{
MagPIE object of the IEA Ammonia data disaggregated to country level.
}
\description{
Convert IEA Ammonia Technology Roadmap 2021 “Fig 2.9 Ammonia production by process route and
scenario in major ammonia producing regions” data to ISO country level.
}
\examples{
\dontrun{
  a <- convertIEA_Ammonia(x)
}

}
\author{
Qianzhi Zhang
}
