% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_Petrochem.R
\name{convertIEA_Petrochem}
\alias{convertIEA_Petrochem}
\title{Convert IEA_Petrochem}
\usage{
convertIEA_Petrochem(x)
}
\arguments{
\item{x}{MAgPIE object containing IEA Ammonia data region resolution}
}
\value{
MAgPIE object of the IEA Ammonia data disaggregated to country level
}
\description{
Convert IEA The Future of  Petrochemicals 2018 “Fig 4.1 Petrochem Production”, "Fig A.1 Petrochem Prod Region",
"Fig 4.5 Petrochem Feedstock","Fig 4.9 Petro Prod Route RTS","Fig 5.10 Petro Prod Route CTS" data as magclass object
}
\examples{
\dontrun{
a <- convertIEA_Petrochem(x)
}

}
\author{
Qianzhi
}
