% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_PetrochemEI.R
\name{convertIEA_PetrochemEI}
\alias{convertIEA_PetrochemEI}
\title{Convert IEA_PetrochemEI}
\usage{
convertIEA_PetrochemEI(x)
}
\arguments{
\item{x}{MAgPIE object containing IEA_PetrochemEI region resolution}
}
\value{
MAgPIE object of the IEA_PetrochemEI data disaggregated to country level
}
\description{
Convert IEA Chemical and Petrochemical Sector 2009 “Table 12. Petro Regional Coef” data as magclass object
}
\examples{
\dontrun{
a <- convertIEA_PetrochemEI(x)
}

}
\author{
Qianzhi Zhang
}
