% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIFA_Chem.R
\name{convertIFA_Chem}
\alias{convertIFA_Chem}
\title{Convert IFA_Chem}
\usage{
convertIFA_Chem(x)
}
\arguments{
\item{x}{MagPIE object containing IFA data at region resolution.}
}
\value{
MagPIE object of the IFA data disaggregated to country level.
}
\description{
Convert IFA_Chem data to ISO country level.
}
\examples{
\dontrun{
  a <- convertIFA_Chem(x)
}

}
\author{
Qianzhi Zhang
}
