% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIFA_ChemAppend.R
\name{convertIFA_ChemAppend}
\alias{convertIFA_ChemAppend}
\title{Convert IFA_Chem}
\usage{
convertIFA_ChemAppend(x)
}
\arguments{
\item{x}{MagPIE object containing IFA data at regional resolution.}
}
\value{
MagPIE object of the IFA data disaggregated to country level.
}
\description{
Convert IFA_Chem data to ISO country level.
}
\examples{
\dontrun{
  a <- convertIFA_ChemAppend(x)
}

}
\author{
Qianzhi Zhang
}
