% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIHS_Meth.R
\name{convertIHS_Meth}
\alias{convertIHS_Meth}
\title{Convert IHS_Meth}
\usage{
convertIHS_Meth(x)
}
\arguments{
\item{x}{MAgPIE object containing IHS_Meth region resolution}
}
\value{
MAgPIE object of the IHS_Meth data disaggregated to country level
}
\description{
Convert 9TH RUSSIA & CIS OIL & GAS EXECUTIVE SUMMIT 2019 METHANOL INS.xlsx data to ISO country level.
}
\examples{
\dontrun{
a <- convertIHS_Meth(x)
}

}
\author{
Qianzhi Zhang
}
