% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMMSA_Methanol.R
\name{convertMMSA_Methanol}
\alias{convertMMSA_Methanol}
\title{Convert MMSA_Methanol}
\usage{
convertMMSA_Methanol(x)
}
\value{
magpie object of the MMSA_Methanol data
}
\description{
Convert MMSA(Methanol Market Services Asia) Global Methanol Outlook 2023 Growth and Decarbonization.xlsx data to ISO country level.
}
\examples{
\dontrun{
a <- convertMMSA_Methanol(x)
}

}
\author{
Qianzhi Zhang
}
