% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOECD_Plastic.R
\name{convertOECD_Plastic}
\alias{convertOECD_Plastic}
\title{Convert OECD Global Plastics Outlook Data to a MagPIE Object}
\usage{
convertOECD_Plastic(x)
}
\arguments{
\item{x}{MagPIE object containing OECD_Plastic data at regional resolution.}
}
\value{
MagPIE object of the OECD_Plastic data disaggregated to country level.
}
\description{
Convert OECD Global Plastics Outlook .xlsx data into a MagPIE object
aggregated to ISO country level
}
\examples{
\dontrun{
  a <- convertOECD_Plastic(x)
}

}
\author{
Qianzhi Zhang
}
