% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPostedBuiltLifespan.R
\name{convertPostedBuiltLifespan}
\alias{convertPostedBuiltLifespan}
\title{convert data from PostedLifetimes}
\usage{
convertPostedBuiltLifespan(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
convert data from PostedLifetimes
}
\author{
Bennet Weiss
}
