% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRMI_China.R
\name{convertRMI_China}
\alias{convertRMI_China}
\title{Convert RMI_China}
\usage{
convertRMI_China(x)
}
\arguments{
\item{x}{MAgPIE object containing RMI_China data region resolution}
}
\value{
MAgPIE object of the RMI_China data disaggregated to country level
}
\description{
Convert RMI (Rocky Mountain Institute) Transforming China’s Chemicals Industry Pathways and Outlook under the Carbon Neutrality Goal 2022.xlsx
"ES1-3 China Chemical Demand", "ES29 China Chemical Structure" data to ISO country level.
}
\examples{
\dontrun{
a <- convertRMI_China(x)
}

}
\author{
Qianzhi Zhang
}
