% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUNComtrade.R
\name{convertUNComtrade}
\alias{convertUNComtrade}
\title{Convert trade data from UN Comtrade.}
\usage{
convertUNComtrade(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert trade data from UN Comtrade.
}
\author{
Bennet Weiss
}
