% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUSGSDS140.R
\name{convertUSGSDS140}
\alias{convertUSGSDS140}
\title{Convert trade data from USGS DS140.}
\usage{
convertUSGSDS140(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert trade data from USGS DS140.
}
\author{
Bennet Weiss
}
