% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSIMSON.R
\name{fullSIMSON}
\alias{fullSIMSON}
\title{Load all SIMSON data}
\usage{
fullSIMSON(scenario = "SSP2", gdp_per_capita = FALSE)
}
\description{
Function that produces the complete regional data sets required for the 
SIMSON model.
}
\examples{
\dontrun{
retrieveData('SIMSON')
fullSIMSON()
}

}
\seealso{
\code{\link[madrat]{readSource}}, \code{\link[madrat]{getCalculations}}, 
\code{\link[madrat]{calcOutput}}, \code{\link[mrindustry]{calcSteelProduction}}
}
\author{
Merlin HOSAK
}
