% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGNR.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize strings for comparison.}
\usage{
normalize_text(strings, lookup)
}
\arguments{
\item{strings}{Character vector to normalize.}

\item{lookup}{Named character vector mapping originals to replacements.}
}
\value{
Character vector of normalized strings.
}
\description{
Normalize strings for comparison.
}
