% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODYM_RECC.R
\name{readODYM_RECC}
\alias{readODYM_RECC}
\alias{calcODYM_RECC}
\title{Read ODYM_RECC data from the SHAPE Project}
\usage{
readODYM_RECC(subtype)

calcODYM_RECC(subtype, smooth = TRUE)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{'REMIND_industry_trends'}: Trends in per-capita production of industry
subsectors \code{cement}, \code{chemicals}, \code{steel_primary}, \code{steel_secondary},
and \code{otherInd}.  Trends for \code{chemicals} and \code{otherInd} are averages of
the other three trends, which are provided by NTNU.
}}

\item{smooth}{Smooth REMIND_industry_trends (default) or not.}
}
\value{
A \code{\link[magclass:magclass]{magpie}} object.
}
\description{
Read ODYM_RECC data from the SHAPE Project
}
\author{
Michaja Pehl
}
