% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readADVANCE_WP2.R
\name{readADVANCE_WP2}
\alias{readADVANCE_WP2}
\title{Read ADVANCE WP2 Data}
\usage{
readADVANCE_WP2(subtype)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{clinker-to-cement-ratio} for the clinker-to-cement ratios from figure 21
of Edelenbosch, O. \emph{Enhancing the representation of energy demand
developments in IAM models - A Modeling Guide for the Cement Industry}
(2015) (zotero://select/items/JP8X2QFK)
}}
}
\value{
A \code{\link[magclass:magclass-package]{magpie}} object.
}
\description{
Read ADVANCE WP2 Data
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}, \code{\link[=convertADVANCE_WP2]{convertADVANCE_WP2()}}
}
\author{
Michaja Pehl
}
