% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndrew2019.R
\name{readAndrew2019}
\alias{readAndrew2019}
\title{Read Data from Andrew's 2019 paper}
\usage{
readAndrew2019()
}
\description{
https://zenodo.org/records/11207133
Last version update 17.05.2024
Dataset update from:
Andrew, R.M., 2019. Global CO2 emissions from cement production, 1928–2018.
Earth System Science Data 11, 1675–1710. https://doi.org/10.5194/essd-11-1675-2019.
}
\author{
Bennet Weiss
}
