% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChathamHouse.R
\name{readChathamHouse}
\alias{readChathamHouse}
\title{Read cement and clinker trade data from Chatham House. Get net trade (export).}
\usage{
readChathamHouse(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement" or "clinker"}
}
\description{
Chatham House (2024), ‘resourcetrade.earth’, https://resourcetrade.earth/.
Accessed: 29.04.2025.
Data based on UN Comtrade.
}
\author{
Bennet Weiss.
}
