% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChina_CNBS.R
\name{readChina_CNBS}
\alias{readChina_CNBS}
\title{Read China_CNBS Data into a magpie Object}
\usage{
readChina_CNBS()
}
\value{
magpie object of the China_CNBS Data
}
\description{
This function reads China CNBS data on plastics EoL treatment in China.
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "China_CNBS")
}
}
\author{
Leonie Schweiger
}
