% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChinaBaogao.R
\name{readChinaBaogao}
\alias{readChinaBaogao}
\title{Read ChinaBaogao}
\usage{
readChinaBaogao()
}
\value{
magpie object of the ChinaBaogao data
}
\description{
Read-in ChinaBaogao Methanol 2023 .xlsx file as
a magclass object.
}
\details{
ChinaBaogao is a Chinese business website that publishes industry analyses.
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "ChinaBaogao")
}
}
\author{
Qianzhi Zhang
}
