% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGNR.R
\name{readGNR}
\alias{readGNR}
\title{Read data from GNR
Getting the Numbers Right (GNR) is a Project of the Global Cement and Concrete Association (GCCA)
[GNR](https://gccassociation.org/sustainability-innovation/gnr-gcca-in-numbers/)
Data not publically available. Data received from Abhishek Shukla, Sustainability Program Manager at GCCA.
Data received on 21.01.2025, personal communication.}
\usage{
readGNR(subtype = "clinker_ratio")
}
\arguments{
\item{subtype}{Variable to be read in. Currently, only "clinker_ratio" is supported.}
}
\description{
Read data from GNR
Getting the Numbers Right (GNR) is a Project of the Global Cement and Concrete Association (GCCA)
[GNR](https://gccassociation.org/sustainability-innovation/gnr-gcca-in-numbers/)
Data not publically available. Data received from Abhishek Shukla, Sustainability Program Manager at GCCA.
Data received on 21.01.2025, personal communication.
}
\author{
Bennet Weiss
}
