% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_Ammonia.R
\name{readIEA_Ammonia}
\alias{readIEA_Ammonia}
\title{Read IEA_Ammonia}
\usage{
readIEA_Ammonia(subtype)
}
\arguments{
\item{subtype}{Different scenarios of Ammonia data that should be read. Available types are:
\itemize{
  \item BaseYear_2020: Base year data in 2020
  \item STEPS_2050: IEA STEPS scenario in 2050
  \item SDS_2050: IEA SDS scenario in 2050
}}
}
\value{
Magpie object of the IEA Ammonia data.
}
\description{
Read-in IEA Ammonia Technology Roadmap 2021 Fig 2.9 Ammonia production by process route and
scenario in major ammonia producing regions data as a magclass object.
}
\seealso{
[readSource()]
\dontrun{
  a <- readSource(type = "IEA_Ammonia", subtype = "BaseYear_2020")
}
}
\author{
Qianzhi Zhang
}
