% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_PetrochemEI.R
\name{readIEA_PetrochemEI}
\alias{readIEA_PetrochemEI}
\title{Read-in specific energy consumption (SEC) for the production of key chemicals (2006)
from IEA Information Paper Chemical and Petrochemical Sector 2009 (Table 12)}
\usage{
readIEA_PetrochemEI()
}
\description{
Read-in specific energy consumption (SEC) for the production of key chemicals (2006)
from IEA Information Paper Chemical and Petrochemical Sector 2009 (Table 12)
}
\author{
Qianzhi Zhang
}
