% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFA_Chem.R
\name{readIFA_Chem}
\alias{readIFA_Chem}
\title{Read IFA}
\usage{
readIFA_Chem(subtype)
}
\arguments{
\item{subtype[1]}{Type of IFA product that should be read. Available types are:
\itemize{ \item ammonia
\item urea}

#' @param subtype[2] Different types of IFA data sheets that should be read. Available types are:
\itemize{ \item statistics: including production, export, import, consumption
\item capacities: including capacities}

#' @param subtype[3] Characteristics of different products that should be read. Available types are:
\itemize{ \item consumption
\item production
\item export
\item import
\item capacities}}
}
\value{
magpie object of the IFA data
}
\description{
Read-in IFA (International Fertilizer Association) data .xlsx file as
magclass object
}
\seealso{
[readSource()]
\dontrun{
a <- readSource(type = "IFA_Chem", subtype = "consumption")
}
}
\author{
Qianzhi Zhang
}
