% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFA_ChemAppend.R
\name{readIFA_ChemAppend}
\alias{readIFA_ChemAppend}
\title{Read IFA}
\usage{
readIFA_ChemAppend(subtype)
}
\arguments{
\item{subtype}{Character string indicating the IFA product and data type to read.
Available combinations include:
\itemize{
  \item AN_statistics_production
  \item AS_statistics_production
  \item CAN_statistics_production
  \item AN_capacities_capacities, AS_capacities_capacities, UAN_capacities_capacities
}}
}
\value{
Magpie object of the IFA data.
}
\description{
Read-in IFA (International fertilizer Association) data .xlsx file containing
production volumes and/or capacities for ammonium nitrate (AN), ammonium sulphate (AS),
calcium ammonium nitrate (CAN) and urea ammonium nitrate (UAN) as a magclass object.
}
\seealso{
[readSource()]

\dontrun{
  a <- readSource(type = "IFA_ChemAppend", subtype = "AN_statistics_production")
}
}
\author{
Qianzhi Zhang
}
