% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMagpieFertilizer.R
\name{readMagpieFertilizer}
\alias{readMagpieFertilizer}
\title{readMagpieFertilizer}
\usage{
readMagpieFertilizer()
}
\description{
Read function for fertilizer data coming from the MAgPIE model.
}
\seealso{
[readSource()]
\dontrun{
  a <- readSource(type = "MagpieFertilizer")
}
}
\author{
Leonie Schweiger
}
