% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMMSA_Methanol.R
\name{readMMSA_Methanol}
\alias{readMMSA_Methanol}
\title{Read MMSA_Methanol}
\usage{
readMMSA_Methanol()
}
\description{
Read-in MMSA Global Methanol Outlook 2023 Growth and Decarbonization.xlsx sheets as a magclass object.
Data contains regional methanol capacities and demands. The data is merged from different figures covering different regions.
}
