% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOECD_GDP.R
\name{readOECD_GDP}
\alias{readOECD_GDP}
\title{Read OECD GDP}
\usage{
readOECD_GDP(subtype = "gdpPC")
}
\arguments{
\item{subtype}{Specific dataset used by that source}

\item{version}{Version of the dataset to read}
}
\description{
Read OECD GDP per capita data from 1500-2016, given in 2011 USD.
}
\author{
Merlin Jo Hosak
}
