% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRMI_China.R
\name{readRMI_China}
\alias{readRMI_China}
\title{Read RMI_China}
\usage{
readRMI_China(subtype)
}
\arguments{
\item{subtype[1]}{Type of RMI_China data sheet to read. Available types are:
\itemize{
  \item ChemDemand: ES1-3 China Chemical Demand
  \item ChemStructure: ES10 China Chemical Structure
}}

\item{subtype[2]}{The specific product from the RMI_China data to read. Available types are:
\itemize{
  \item Ammonia
  \item Methanol
  \item Ethylene
}}
}
\value{
MagPIE object of the RMI_China data.
}
\description{
Read-in RMI (Rocky Mountain Institute) "Transforming China Chemicals Industry Pathways and Outlook under the Carbon Neutrality Goal 2022.xlsx"
data from either "ES1-3 China Chemical Demand" or "ES29 China Chemical Structure" sheets as a magclass object.
Data contains chemical demand projections 2020-2050 for ammonia, methanol and ethylene and the feedstock structure for the production of ammonia, methanol and ethylene
in the zero-carbon scenario.
}
\seealso{
[readSource()]
\dontrun{
  a <- readSource(type = "RMI_China", subtype = "ChemDemand_Ammonia")
}
}
\author{
Qianzhi Zhang
}
