% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNComtrade.R
\name{readUNComtrade}
\alias{readUNComtrade}
\title{Read cement and clinker trade data from UN Comtrade. Get net trade (export).}
\usage{
readUNComtrade(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement" or "clinker"}
}
\description{
United Nations Statistics Division. (n.d.). UN Comtrade Database.
Retrieved May 8, 2025, from https://comtradeplus.un.org/
}
\author{
Bennet Weiss.
}
