% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNCTAD.R
\name{readUNCTAD}
\alias{readUNCTAD}
\title{Read UNCTAD}
\usage{
readUNCTAD()
}
\value{
magpie object of the CUNCTAD data
}
\description{
Read-in US_PlasticsTradebyPartner file as
a magclass object.
}
\details{
UNCTAD is United Nations Conference on Trade and Development.
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "UNCTAD")
}
}
\author{
Qianzhi Zhang
}
