% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUSGSDS140.R
\name{readUSGSDS140}
\alias{readUSGSDS140}
\title{Read cement statistic data sheet from USGS.
It combines historic data from USGS sources. Supply-Demand Statistics.
Get net trade (= exports) of cement (which includes clinker trade).}
\usage{
readUSGSDS140()
}
\description{
U.S. Geological Survey, 2020, Cement statistics, in Kelly, T.D., and Matos, G.R., comps.,
Historical statistics for mineral and material commodities in the United States:
U.S. Geological Survey Data Series 140
accessed 12.02.2025, at
https://www.usgs.gov/centers/national-minerals-information-center/
historical-statistics-mineral-and-material-commodities.
}
\author{
Bennet Weiss.
}
