% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustry_Value_Added.R
\name{toolGetCementConvergenceParameters}
\alias{toolGetCementConvergenceParameters}
\title{convergence year and level (relative to global average) to which per-capita
cement demand converges (Michaja Pehl)}
\usage{
toolGetCementConvergenceParameters()
}
\description{
convergence year and level (relative to global average) to which per-capita
cement demand converges (Michaja Pehl)
}
