% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSteel_Projections.R
\name{toolGetExpertGuessSteelProduction}
\alias{toolGetExpertGuessSteelProduction}
\title{Steel production estimates (Michaja Pehl)}
\usage{
toolGetExpertGuessSteelProduction()
}
\description{
Steel production estimates (Michaja Pehl)
}
