% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetRegionCountries.R
\name{toolGetRegionMask}
\alias{toolGetRegionMask}
\title{Get region mask for Magpie object that selects countries of a region based on a region mapping.}
\usage{
toolGetRegionMask(x, region, regionmapping)
}
\arguments{
\item{x}{Magpie object}

\item{region}{string or vector of strings to identify the region}

\item{regionmapping}{mapping of regions to countries}
}
\description{
Get region mask for Magpie object that selects countries of a region based on a region mapping.
}
\author{
Bennet Weiss
}
