% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcindustry_max_secondary_steel_share.R
\name{toolGetSecondarySteelShare}
\alias{toolGetSecondarySteelShare}
\title{Maximum share of secondary steel production in total steel production and
years between which a linear convergence from historic to target shares is
to be applied. (Michaja Pehl)}
\usage{
toolGetSecondarySteelShare()
}
\description{
Maximum share of secondary steel production in total steel production and
years between which a linear convergence from historic to target shares is
to be applied. (Michaja Pehl)
}
