% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolInterpolate2D.R
\name{toolInterpolate2D}
\alias{toolInterpolate2D}
\title{Interpolate missing values in a 2D magpie object}
\usage{
toolInterpolate2D(x, method = "linear", ...)
}
\arguments{
\item{method}{3 interpolation methods are supported: 'linear', 'spline', and 'constant' (last observation carried forward).}
}
\description{
This function interpolates missing values (NAs) in a 2D magpie 
object along the time dimension (years).
}
\author{
Merlin Jo Hosak
}
