% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMerge2D.R
\name{toolMerge2D}
\alias{toolMerge2D}
\title{Merge two 2D magpie objects}
\usage{
toolMerge2D(a, b, fill = NA)
}
\description{
Perform an outer merge on two 2D magpie objects (regions x years).
}
\author{
Merlin Jo Hosak
}
