% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNAHandling.R
\name{toolRemoveNA}
\alias{toolRemoveNA}
\title{Removes NA values from a Magpie object. Only works if all regions show the same temporal pattern of data scarcity.}
\usage{
toolRemoveNA(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Removes NA values from a Magpie object. Only works if all regions show the same temporal pattern of data scarcity.
}
\author{
Bennet Weiss
}
