% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyBalancesOutputToIndustry.R
\name{calcEnergyBalancesOutputToIndustry}
\alias{calcEnergyBalancesOutputToIndustry}
\title{Calculate Energy Balances Output to Industry
Additional corrections are applied to the IEA data in [`mrindustry::tool_fix_IEA_data_for_Industry_subsectors`].}
\usage{
calcEnergyBalancesOutputToIndustry()
}
\description{
Calculate Energy Balances Output to Industry
Additional corrections are applied to the IEA data in [`mrindustry::tool_fix_IEA_data_for_Industry_subsectors`].
}
\author{
Michaja Pehl, Falk Benke
}
