% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeDemandIndustry.R
\name{calcFeDemandIndustry}
\alias{calcFeDemandIndustry}
\title{Calculates FE demand in industry as REMIND variables}
\usage{
calcFeDemandIndustry(
  scenarios,
  use_ODYM_RECC = FALSE,
  last_empirical_year = 2020
)
}
\arguments{
\item{scenarios}{Vector of strings designating the scenario. !!Currently it acts as a filter only, with the actual
scenarios computed within the function hard-coded into remind_scenarios.}

\item{use_ODYM_RECC}{per-capita pathways for \code{SDP_xx} scenarios?  (Defaults
to \code{FALSE}.)}

\item{last_empirical_year}{Last year for which empirical data is available.
Defaults to 2020.}
}
\description{
Calculates FE demand in industry as REMIND variables
}
\author{
Michaja Pehl
}
