% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndustry_Value_Added.R
\name{calcIndustry_Value_Added}
\alias{calcIndustry_Value_Added}
\title{Function for calculating industry activity trajectories.}
\usage{
calcIndustry_Value_Added(
  subtype = "physical",
  match.steel.historic.values = TRUE,
  match.steel.estimates = "none",
  save.plots = NULL,
  China_Production = NULL
)
}
\arguments{
\item{subtype}{One of
\itemize{
\item \code{physical} Returns physical production trajectories for cement.
\item \code{economic} Returns value added trajectories for all subsectors.
}}

\item{match.steel.historic.values}{Should steel production trajectories match
historic values?}

\item{match.steel.estimates}{Should steel production trajectories match
exogenous estimates?  \code{NULL} or one of
\itemize{
\item \code{IEA_ETP} IEA 2017 Energy Transition Pathways steel production totals for
OECD and Non-OECD countries from the \emph{Reference Technologies Scenario}
until 2060, and original growth rates after that.
}}

\item{save.plots}{\code{NULL} (default) if no plots are saved, or the path to save
directories to.}

\item{China_Production}{A data frame with columns \code{period} and
\code{total.production} prescribing total production for China to have,
disregarding results from the stock saturation model.}
}
\value{
A list with a \code{\link[magclass:magclass-package]{magpie}} object \code{x}, \code{weight},
\code{unit}, \code{description}, \code{min}, and \code{max}.
}
\description{
Function for calculating industry activity trajectories.
}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Michaja Pehl
}
