% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAndrew2019.R
\name{convertAndrew2019}
\alias{convertAndrew2019}
\title{Convert data from Andrew 2019.}
\usage{
convertAndrew2019(x, subtype)
}
\arguments{
\item{x}{Magpie object}

\item{subtype}{Material subtype. Can be "cement" or "clinker".}
}
\description{
Convert data from Andrew 2019.
}
\author{
Bennet Weiss
}
