% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertChinaBaogao.R
\name{convertChinaBaogao}
\alias{convertChinaBaogao}
\title{Convert ChinaBaogao}
\usage{
convertChinaBaogao(x)
}
\arguments{
\item{x}{MAgPIE object containing ChinaBaogao Methanol region resolution}
}
\value{
MAgPIE object of the ChinaBaogao Methanol data disaggregated to country level
}
\description{
Convert ChinaBaogao Methanol 2023.xlsx data to ISO country level.
}
\examples{
\dontrun{
a <- convertChinaBaogao(x)
}

}
\author{
Qianzhi Zhang
}
