% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_Petrochem.R
\name{convertIEA_Petrochem}
\alias{convertIEA_Petrochem}
\title{Convert IEA_Petrochem}
\usage{
convertIEA_Petrochem(x)
}
\arguments{
\item{x}{MagPIE object containing IEA Petrochem data at regional resolution.}
}
\value{
MagPIE object of the IEA Petrochem data disaggregated to country level.
}
\description{
Convert IEA The Future of Petrochemicals 2018 data (including several figures such as
"Fig 4.1 Petrochem Production", "Fig A.1 Petrochem Prod Region", "Fig 4.5 Petrochem Feedstock",
"Fig 4.9 Petro Prod Route RTS", "Fig 5.10 Petro Prod Route CTS") into a MagPIE object
aggregated to ISO country level.
}
\examples{
\dontrun{
  a <- convertIEA_Petrochem(x)
}

}
\author{
Qianzhi
}
