% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPostedBuiltLifespan.R
\name{convertPostedBuiltLifespan}
\alias{convertPostedBuiltLifespan}
\title{Convert data from PostedBuiltLifespan. Lifetimes are assigned to region based on regionmapping.}
\usage{
convertPostedBuiltLifespan(x)
}
\arguments{
\item{x}{Magpie object}
}
\description{
Convert data from PostedBuiltLifespan. Lifetimes are assigned to region based on regionmapping.
}
\author{
Bennet Weiss
}
