% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrindustry-package.R
\docType{package}
\name{mrindustry-package}
\alias{mrindustry}
\alias{mrindustry-package}
\title{mrindustry: input data generation for the REMIND industry module}
\description{
The mrindustry packages contains data preprocessing for the REMIND model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrindustry}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Jakob Dürrwächter \email{jakobdu@pik-potsdam.de}
}

}
\keyword{internal}
