% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndrew2019.R
\name{readAndrew2019}
\alias{readAndrew2019}
\title{Read Data from Andrew's 2019 paper}
\usage{
readAndrew2019(subtype)
}
\arguments{
\item{subtype}{Material subtype. Can be "cement" or "clinker".}
}
\description{
https://zenodo.org/records/14931651
Last version update 26.02.2025
Dataset update from:
Andrew, R.M., 2019. Global CO2 emissions from cement production, 1928–2018.
Earth System Science Data 11, 1675–1710. https://doi.org/10.5194/essd-11-1675-2019.
}
\author{
Bennet Weiss
}
