% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChinaBaogao.R
\name{readChinaBaogao}
\alias{readChinaBaogao}
\title{Read ChinaBaogao}
\usage{
readChinaBaogao()
}
\value{
magpie object of the China Methanol data
}
\description{
Read-in ChinaBaogao(This is a website containing research reports on different Chinese industries) Methanol 2023.xlsx file as
magclass object
}
\seealso{
[readSource()]
\dontrun{
a <- readSource(type = "ChinaBaogao")
}
}
\author{
Qianzhi Zhang
}
