% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFA_Chem.R
\name{readIFA_Chem}
\alias{readIFA_Chem}
\title{Read IFA}
\usage{
readIFA_Chem(subtype)
}
\arguments{
\item{subtype}{Character string specifying the type of IFA product data to read.
Available types are:
\itemize{
  \item For product: ammonia, urea
  \item For data sheet: statistics, capacities
  \item For product characteristics: consumption, production, export, import, capacities
}}
}
\value{
magpie object of the IFA data
}
\description{
Read-in IFA (International fertilizer Association) data .xlsx file containing 
production, consumption, export and import volumes as well as capacities for 
ammonia and urea as a magclass object.
}
\seealso{
[readSource()]

\dontrun{
a <- readSource(type = "IFA_Chem", subtype = "ammonia_statistics_production")
}
}
\author{
Qianzhi Zhang
}
