% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIHS_Meth.R
\name{readIHS_Meth}
\alias{readIHS_Meth}
\title{Read IHS_Meth}
\usage{
readIHS_Meth(subtype)
}
\arguments{
\item{subtype[1]}{Type of Methanol data that should be read. Available types are:
\itemize{ \item 2010-2020: Interpolated 2018 data to 2010-2020 using data on total global methanol production
\item 2018}}
}
\value{
magpie object of the IHS_Meth data
}
\description{
Read-in 9TH RUSSIA & CIS OIL & GAS EXECUTIVE SUMMIT 2019 METHANOL INS .xlsx file as
magclass object
}
\seealso{
[readSource()]
\dontrun{
a <- readSource(type = "IHS_Meth", subtype = "Production_2018")
}
}
\author{
Qianzhi Zhang
}
