% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMMSA_Methanol.R
\name{readMMSA_Methanol}
\alias{readMMSA_Methanol}
\title{Read MMSA_Methanol}
\usage{
readMMSA_Methanol()
}
\value{
magpie object of the MMSA_Methanol data
}
\description{
Read-in MMSA(Methanol Market Services Asia) Global Methanol Outlook 2023 Growth and Decarbonization.xlsx file as
magclass object
}
\seealso{
[readSource()]
\dontrun{
a <- readSource(type = "MMSA_Methanol")
}
}
\author{
Qianzhi Zhang
}
