% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPostedBuiltLifespan.R
\name{readPostedBuiltLifespan}
\alias{readPostedBuiltLifespan}
\title{Read Data as prepared by Posted}
\usage{
readPostedBuiltLifespan()
}
\description{
https://github.com/PhilippVerpoort/posted
Data sources can be found in Posted.
}
\author{
Bennet Weiss
}
