% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRMI_China.R
\name{readRMI_China}
\alias{readRMI_China}
\title{Read RMI_China}
\usage{
readRMI_China(subtype)
}
\arguments{
\item{subtype[1]}{Type of RMI_China data sheets that should be read. Available types are:
\itemize{ \item ChemDemand: ES1-3 China Chemical Demand
\item ChemStructure:ES29 China Chemical Structure }

#' @param subtype[2] Different products from RMI_China data that should be read. Available types are:
\itemize{ \item Ammonia
\item Methanol
\item Ethylene}}
}
\value{
magpie object of the RMI_China data
}
\description{
Read-in RMI (Rocky Mountain Institute) Transforming China Chemicals Industry Pathways and Outlook under the Carbon Neutrality Goal 2022.xlsx
"ES1-3 China Chemical Demand", "ES29 China Chemical Structure" files as
magclass object
}
\seealso{
[readSource()]
\dontrun{
a <- readSource(type = "RMI_China", subtype = "ChemDemand_Ammonia")
}
}
\author{
Qianzhi Zhang
}
