% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetRegionCountries.R
\name{toolGetRegionCountries}
\alias{toolGetRegionCountries}
\title{Retrieve countries from a Magpie object that belong to a region based on a regionmapping.}
\usage{
toolGetRegionCountries(x, region, regionmapping)
}
\arguments{
\item{x}{Magpie object}

\item{region}{string or vector of strings to identify the region}

\item{regionmapping}{mapping of regions to countries}
}
\description{
Retrieve countries from a Magpie object that belong to a region based on a regionmapping.
}
\author{
Bennet Weiss
}
